requirements.txt
torch
torchvision
Pillow
matplotlib
streamlit
pandas
numpy
